fereatra de confirmare pentru stergere

+++ resources/views/admin/indrumatori/index.blade.php
@@ -3,7 +3,7 @@
 
 @section('content')
 
-    <h1>Utilizatori</h1>
+    <h1>Elevi</h1>
 
     <table class="table table-striped">
         <thead>
@@ -23,16 +23,16 @@
             @foreach($users as $user)
                 <tr>
                     <th scope="row">{{$user->id}}</th>
-                    <th scope="row">{{$user->name}}</th>
-                    <th scope="row">{{$user->email}}</th>
-                    <th scope="row">{{$user->judete->nume}}/{{$user->localitati->nume}}</th>
-                    <th scope="row">{{(!empty($user->school))?$user->school->name:""}}</th>
-                    <th scope="row">{{(!empty($user->prof))?$user->prof->name:""}}</th>
-                    <th scope="row">{{title_case($user->role->name)}}</th>
-                    <th scope="row">{{($user->active==1)?'Da':'Nu'}}</th>
-                    <th scope="row"><a href="{{route("admin.indrumatori.edit", $user->id)}}" style="text-decoration: none"> <img src="/img/edit.png" height="25"></a></th>
-                    <th scope="row">
-                        {!! Form::open(['method'=>'DELETE','action'=>['AdminIndrumatoriController@destroy',$user->id]]) !!}
+                    <td>{{$user->name}}</td>
+                    <td>{{$user->email}}</td>
+                    <td>{{$user->judete->nume}}/{{$user->localitati->nume}}</td>
+                    <td>{{(!empty($user->school))?$user->school->name:""}}</td>
+                    <td>{{(!empty($user->prof))?$user->prof->name:""}}</td>
+                    <td>{{title_case($user->role->name)}}</td>
+                    <td>{{($user->active==1)?'Da':'Nu'}}</td>
+                    <td><a href="{{route("admin.indrumatori.edit", $user->id)}}" style="text-decoration: none"> <img src="/img/edit.png" height="25"></a></td>
+                    <td>
+                        {!! Form::open(['method'=>'DELETE','action'=>['AdminIndrumatoriController@destroy',$user->id], 'onsubmit' => 'return ConfirmDialog("Sigur vreți să eliminați utilizatorul?")']) !!}
 
                         <div class="form-group">
                             {{--{!! Form::submit('', ['style'=>'background: url("/img/delete.png") no-repeat scroll 0 0 transparent;color: #000000;cursor: pointer;font-weight: bold;height: 20px;padding-bottom: 2px;width: 75px;']) !!}--}}
@@ -43,11 +43,25 @@
                         {!! Form::close() !!}
 
 
-                    </th>
+                    </td>
                 </tr>
             @endforeach
         @endif
         </tbody>
     </table>
 
+@endsection
+
+@section('footer')
+    <script>
+
+        function ConfirmDialog(message){
+            var x = confirm(message);
+            if (x)
+                return true;
+            else
+                return false;
+        };
+
+    </script>
 @endsection
