+++ resources/views/admin/users/index.blade.php
@@ -30,17 +30,18 @@
                     <th scope="row">{{(!empty($user->prof))?$user->prof->name:""}}</th>
                     <th scope="row">{{title_case($user->role->name)}}</th>
                     <th scope="row">{{($user->active==1)?'Da':'Nu'}}</th>
-                    <th scope="row"><a href="{{route("admin.users.edit", $user->id)}}" style="text-decoration: none"> <img src="/img/edit.png" height="25"></a></th>
+                    <th scope="row"><a href="{{route("admin.users.edit", $user->id)}}" style="text-decoration: none">
+                            <img src="/img/edit.png" height="25"></a></th>
                     <th scope="row">
-                        {!! Form::open(['method'=>'DELETE','action'=>['AdminUsersController@destroy',$user->id]]) !!}
+                        {!! Form::open(['method'=>'DELETE','action'=>['AdminUsersController@destroy',$user->id], 'onsubmit' => 'return ConfirmDialog("Sigur vreți să eliminați utilizatorul?")']) !!}
 
-                            <div class="form-group">
-                                {{--{!! Form::submit('', ['style'=>'background: url("/img/delete.png") no-repeat scroll 0 0 transparent;color: #000000;cursor: pointer;font-weight: bold;height: 20px;padding-bottom: 2px;width: 75px;']) !!}--}}
-                                {{--<input type="image" src="/img/delete.png" height="20" alt="Submit" />--}}
-                                {!! Form::image('img/delete.png','success', array( 'height' => 25 ))  !!}
-                            </div>
+                        <div class="form-group">
+                            {{--{!! Form::submit('', ['style'=>'background: url("/img/delete.png") no-repeat scroll 0 0 transparent;color: #000000;cursor: pointer;font-weight: bold;height: 20px;padding-bottom: 2px;width: 75px;']) !!}--}}
+                            {{--<input type="image" src="/img/delete.png" height="20" alt="Submit" />--}}
+                            {!! Form::image('img/delete.png','success', array( 'height' => 25 ))  !!}
+                        </div>
 
-                            {!! Form::close() !!}
+                        {!! Form::close() !!}
 
 
                     </th>
@@ -50,4 +51,18 @@
         </tbody>
     </table>
 
+@endsection
+
+@section('footer')
+    <script>
+
+        function ConfirmDialog(message){
+            var x = confirm(message);
+            if (x)
+                return true;
+            else
+                return false;
+        };
+
+    </script>
 @endsection
